requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
email-validator
